refactor:change Item.id to long

+++ src/main/java/ru/yandex/practicum/filmorate/model/Item.java
@@ -7,5 +7,5 @@ import lombok.*;
 @ToString
 public abstract class Item {
     @EqualsAndHashCode.Exclude
-    protected int id;
+    protected long id;
 }
